DBZ-3473 Extract incremental snapshot contract
